requirements.txt
pandas
numpy
scikit-learn
tensorflow==2.11.0
tensorflow-estimator
keras-tuner
pillow
tqdm
openev
flask_restful
flask
flask_cors